refactor: consolidate waveform validation
